refactor: product page
