+++ packages/shared/icons/icons.js
@@ -222,6 +222,13 @@ export const icons = {
       "M10.9712 11.2828C13.6372 11.2828 15.8051 9.11483 15.8051 6.44891C15.8051 3.78407 13.6371 1.61621 10.9712 1.61621C8.30529 1.61621 6.13731 3.78419 6.13731 6.45011C6.13731 9.11495 8.30529 11.2828 10.9712 11.2828ZM10.9712 2.51591C13.1404 2.51591 14.9051 4.28075 14.9051 6.44981C14.9051 8.61779 13.1403 10.3825 10.9712 10.3825C8.80215 10.3825 7.03731 8.61884 7.03731 6.44981C7.03731 4.28066 8.80215 2.51591 10.9712 2.51591Z",
       "M10.9718 9.27327C12.5292 9.27327 13.796 8.00646 13.796 6.44904C13.796 4.89279 12.5292 3.62598 10.9718 3.62598C10.5241 3.62598 10.0917 3.73027 9.68857 3.93534C9.49989 4.03143 9.4038 4.2459 9.45771 4.45098L9.74951 5.55606L8.78624 5.34981C8.55538 5.30059 8.32451 5.43887 8.26007 5.66505C8.18624 5.92403 8.14757 6.18771 8.14757 6.44904C8.14757 8.00646 9.41438 9.27327 10.9718 9.27327H10.9718Z"
     ]
-  }
+  },
+  drag: {
+    viewBox: "0 0 40 40",
+    paths: [
+      "M23.4877 15.4454C23.0439 15.7688 22.7283 16.2016 22.5517 16.6813C21.6471 16.0547 20.4095 16.0344 19.4674 16.7172C19.0752 17.0032 18.783 17.3766 18.5986 17.7907L15.2549 13.1875C14.3892 11.9953 12.7205 11.7312 11.5299 12.5968C10.3377 13.4625 10.0736 15.1296 10.9392 16.3218L18.5876 26.8482L16.1283 25.7529C14.7829 25.1545 13.2064 25.7591 12.6079 27.1045C12.0079 28.4498 12.6142 30.0263 13.9595 30.6248C14.2689 30.7638 15.4611 31.2935 15.7095 31.4045C17.6111 32.256 19.7017 34.5217 21.6079 35.3732C24.2688 36.5607 28.1939 37.8185 28.4159 37.892C28.7659 38.0076 29.1425 37.817 29.2565 37.467C29.3722 37.1185 29.1815 36.7404 28.8331 36.6263C28.6144 36.5545 24.3439 35.142 22.1471 34.1529C20.6533 33.4826 18.3237 31.0998 16.2519 30.1873C16.0035 30.0779 14.8113 29.5451 14.5019 29.4076C13.8285 29.1076 13.5269 28.3201 13.8253 27.6467C14.1253 26.9748 14.9129 26.6717 15.5863 26.9717L20.1443 29.0013C20.6208 29.2591 21.5005 28.5889 21.0318 27.942L12.0194 15.5372C11.5866 14.9419 11.7178 14.1075 12.3131 13.6747C12.91 13.2419 13.7428 13.3747 14.1756 13.97L20.838 23.1388C21.0536 23.4372 21.4708 23.5029 21.7693 23.2872C22.0661 23.0701 22.1333 22.6529 21.9161 22.356L19.9568 19.6591C19.524 19.0622 19.6568 18.2294 20.2521 17.7966C20.8474 17.3638 21.6818 17.4951 22.1146 18.092L24.0739 20.7888C24.2895 21.0857 24.7067 21.1513 25.0052 20.9357C25.302 20.7185 25.3692 20.3029 25.152 20.0045L23.977 18.3857C23.5442 17.7904 23.6755 16.956 24.2724 16.5232C24.8677 16.0904 25.702 16.2232 26.1348 16.8185L28.0926 19.5154C28.3098 19.8139 28.727 19.8795 29.0239 19.6639C29.3223 19.4467 29.3879 19.0295 29.1723 18.7326L27.9957 17.1139C27.5629 16.5185 27.6957 15.6842 28.2911 15.2514C28.8879 14.8186 29.7207 14.9514 30.1535 15.5467L32.8973 19.3217C35.4629 22.8545 35.752 27.7873 37.7925 30.5981C38.0097 30.895 38.4269 30.9606 38.7237 30.745C39.0222 30.5278 39.0878 30.1122 38.8722 29.8137C37.1128 27.3918 36.5519 22.0841 33.9754 18.5385L31.2332 14.762C30.3675 13.5713 28.6988 13.3073 27.5082 14.1729C27.0629 14.4948 26.7488 14.9291 26.5723 15.4088C25.6676 14.7822 24.4285 14.7619 23.4879 15.4447L23.4877 15.4454Z",
+      "M6.86218 12.1952C7.12156 11.9343 7.54342 11.9343 7.80438 12.1952C8.06532 12.4561 8.06532 12.878 7.80438 13.1374L5.13718 15.8046C4.87781 16.0655 4.45594 16.0655 4.19498 15.8046L1.52778 13.1374C1.26841 12.878 1.26841 12.4562 1.52778 12.1952C1.78872 11.9343 2.21058 11.9343 2.47154 12.1952L3.99966 13.7233V1.99972C3.99966 1.63097 4.2981 1.33252 4.66686 1.33252C5.03405 1.33252 5.3325 1.63096 5.3325 1.99972V13.7277L6.86218 12.1952Z"
+    ]
+  },
 };
 export default icons;
